requirements-dev.txt
asgiref==3.6.0
attrs==23.1.0
bcrypt==4.0.1
blessed==1.20.0
botocore==1.29.99
cached-property==1.5.2
cement==2.8.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
Django==4.2
django-environ==0.10.0
djangorestframework==3.14.0
docopt==0.6.2
gunicorn==20.1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==3.2.0
MarkupSafe==2.1.2
paramiko==3.1.0
pathspec==0.10.1
pycparser==2.21
PyNaCl==1.5.0
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
django-cors-headers==4.1.0
semantic-version==2.8.5
simplejson==3.19.1
six==1.14.0
termcolor==1.1.0
texttable==1.6.7
uritemplate==4.1.1
urllib3==1.26.15
wcwidth==0.1.9
websocket-client==0.59.0
pypdf==3.12.1
sentence-transformers==2.2.2
tiktoken==0.4.0
langchain>=0.0.235
pytest>=7.2.2
openai>=0.27.8
docx2txt==0.8
scipy==1.11.1
nltk==3.8.1
chromadb==0.3.29
transformers>=4.31.0
boto3==1.28.23
botocore==1.31.23